further compartmentalized tests for just create command separate from command runner
